requirements.txt
fastapi
fastapi[standard]
uvicorn
sqlalchemy
psycopg2-binary
pyjwt
pwdlib[argon2]
websockets
python-dotenv
alembic
pydantic
pydantic-settings